move splitPath and joinPath functions to path.ts; add tests for these functions;
